src/requirements.txt
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coloredlogs==15.0.1
humanfriendly==10.0
idna==3.3
protobuf==3.20.0
pycparser==2.21
PyNaCl==1.5.0
requests==2.27.1
tomlkit==0.10.1
urllib3==1.26.9
